I18N-1323 Update Mojito CLI to use OpenAPI spec for rest calls Refactored AI Prompt commands
