+++ cli/src/main/java/com/box/l10n/mojito/cli/command/CreateAIPromptCommand.java
@@ -2,9 +2,10 @@ package com.box.l10n.mojito.cli.command;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import com.box.l10n.mojito.cli.apiclient.AiPromptWsApi;
+import com.box.l10n.mojito.cli.apiclient.ApiException;
 import com.box.l10n.mojito.cli.console.ConsoleWriter;
-import com.box.l10n.mojito.rest.client.AIServiceClient;
-import com.box.l10n.mojito.rest.entity.AIPromptCreateRequest;
+import com.box.l10n.mojito.cli.model.AIPromptCreateRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +21,7 @@ public class CreateAIPromptCommand extends Command {
 
   static Logger logger = LoggerFactory.getLogger(CreateAIPromptCommand.class);
 
-  @Autowired AIServiceClient aiServiceClient;
+  @Autowired AiPromptWsApi aiServiceClient;
 
   @Parameter(
       names = {"--repository-name", "-r"},
@@ -91,7 +92,12 @@ public class CreateAIPromptCommand extends Command {
       throw new CommandException("jsonResponseKey is required when isJsonResponse is true");
     }
     aiPromptCreateRequest.setJsonResponseKey(jsonResponseKey);
-    long promptId = aiServiceClient.createPrompt(aiPromptCreateRequest);
+    long promptId;
+    try {
+      promptId = aiServiceClient.createPrompt(aiPromptCreateRequest);
+    } catch (ApiException e) {
+      throw new CommandException(e.getMessage(), e);
+    }
     consoleWriter.newLine().a("Prompt created with id: " + promptId).println();
   }
 }
