+++ cli/src/main/java/com/box/l10n/mojito/cli/command/CreateAIPromptContextMessageCommand.java
@@ -2,9 +2,10 @@ package com.box.l10n.mojito.cli.command;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import com.box.l10n.mojito.cli.apiclient.AiPromptWsApi;
+import com.box.l10n.mojito.cli.apiclient.ApiException;
 import com.box.l10n.mojito.cli.console.ConsoleWriter;
-import com.box.l10n.mojito.rest.client.AIServiceClient;
-import com.box.l10n.mojito.rest.entity.AIPromptContextMessageCreateRequest;
+import com.box.l10n.mojito.cli.model.AIPromptContextMessageCreateRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,7 +21,7 @@ public class CreateAIPromptContextMessageCommand extends Command {
 
   static Logger logger = LoggerFactory.getLogger(CreateAIPromptContextMessageCommand.class);
 
-  @Autowired AIServiceClient aiServiceClient;
+  @Autowired AiPromptWsApi aiServiceClient;
 
   @Parameter(
       names = {"--content", "-c"},
@@ -61,8 +62,13 @@ public class CreateAIPromptContextMessageCommand extends Command {
     AIPromptContextMessageCreateRequest.setMessageType(messageType);
     AIPromptContextMessageCreateRequest.setAiPromptId(promptId);
     AIPromptContextMessageCreateRequest.setOrderIndex(orderIndex);
-    long contextMessageId =
-        aiServiceClient.createPromptContextMessage(AIPromptContextMessageCreateRequest);
+    long contextMessageId;
+    try {
+      contextMessageId =
+          this.aiServiceClient.createPromptMessage(AIPromptContextMessageCreateRequest);
+    } catch (ApiException e) {
+      throw new CommandException(e.getMessage(), e);
+    }
     consoleWriter
         .newLine()
         .a("Prompt context message created with id: " + contextMessageId)
