+++ cli/src/main/java/com/box/l10n/mojito/cli/command/DeleteAIPromptCommand.java
@@ -2,8 +2,9 @@ package com.box.l10n.mojito.cli.command;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import com.box.l10n.mojito.cli.apiclient.AiPromptWsApi;
+import com.box.l10n.mojito.cli.apiclient.ApiException;
 import com.box.l10n.mojito.cli.console.ConsoleWriter;
-import com.box.l10n.mojito.rest.client.AIServiceClient;
 import org.fusesource.jansi.Ansi;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,7 +21,7 @@ public class DeleteAIPromptCommand extends Command {
 
   static Logger logger = LoggerFactory.getLogger(DeleteAIPromptCommand.class);
 
-  @Autowired AIServiceClient AIServiceClient;
+  @Autowired AiPromptWsApi AIServiceClient;
 
   @Parameter(
       names = {"--prompt-id", "-pi"},
@@ -42,6 +43,10 @@ public class DeleteAIPromptCommand extends Command {
 
   private void deletePrompt() {
     logger.debug("Received request to delete prompt {}", promptId);
-    AIServiceClient.deletePrompt(promptId);
+    try {
+      AIServiceClient.deletePrompt(promptId);
+    } catch (ApiException e) {
+      throw new CommandException(e.getMessage(), e);
+    }
   }
 }
