+++ cli/src/main/java/com/box/l10n/mojito/cli/command/DeleteAIPromptContextMessageCommand.java
@@ -2,8 +2,9 @@ package com.box.l10n.mojito.cli.command;
 
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.Parameters;
+import com.box.l10n.mojito.cli.apiclient.AiPromptWsApi;
+import com.box.l10n.mojito.cli.apiclient.ApiException;
 import com.box.l10n.mojito.cli.console.ConsoleWriter;
-import com.box.l10n.mojito.rest.client.AIServiceClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,7 +20,7 @@ public class DeleteAIPromptContextMessageCommand extends Command {
 
   static Logger logger = LoggerFactory.getLogger(DeleteAIPromptContextMessageCommand.class);
 
-  @Autowired AIServiceClient aiServiceClient;
+  @Autowired AiPromptWsApi aiServiceClient;
 
   @Parameter(
       names = {"--id", "-i"},
@@ -36,7 +37,11 @@ public class DeleteAIPromptContextMessageCommand extends Command {
 
   private void deletePromptContextMessage() {
     logger.debug("Received request to create prompt content message");
-    aiServiceClient.deletePromptContextMessage(id);
+    try {
+      aiServiceClient.deletePromptMessage(id);
+    } catch (ApiException e) {
+      throw new CommandException(e.getMessage(), e);
+    }
     consoleWriter.newLine().a("Prompt context message deleted").println();
   }
 }
